start to decouple ArtemisAuthenticationProvider and UserService
